+++ spec/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/StringFilterSpec.php
@@ -11,7 +11,7 @@ class StringFilterSpec extends ObjectBehavior
 {
     function let(QueryBuilder $queryBuilder)
     {
-        $this->beConstructedWith(['pim_catalog_identifier'], [], ['STARTS WITH', 'ENDS WITH', 'CONTAINS', 'DOES NOT CONTAIN', '=', 'IN']);
+        $this->beConstructedWith(['pim_catalog_identifier'], ['STARTS WITH', 'ENDS WITH', 'CONTAINS', 'DOES NOT CONTAIN', '=', 'IN']);
         $this->setQueryBuilder($queryBuilder);
     }
