+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/Filter/BooleanFilter.php
@@ -15,19 +15,13 @@ use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class BooleanFilter implements AttributeFilterInterface, FieldFilterInterface
+class BooleanFilter extends AbstractFilter implements FieldFilterInterface, AttributeFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedAttributes;
 
     /** @var array */
     protected $supportedFields;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the filter
@@ -49,20 +43,6 @@ class BooleanFilter implements AttributeFilterInterface, FieldFilterInterface
     /**
      * {@inheritdoc}
      */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!($queryBuilder instanceof QueryBuilder)) {
-            throw new \InvalidArgumentException(
-                'Query builder should be an instance of Doctrine\ODM\MongoDB\Query\Builder'
-            );
-        }
-
-        $this->qb = $queryBuilder;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function supportsField($field)
     {
         return in_array($field, $this->supportedFields);
@@ -74,22 +54,6 @@ class BooleanFilter implements AttributeFilterInterface, FieldFilterInterface
     public function supportsAttribute(AttributeInterface $attribute)
     {
         return in_array($attribute->getAttributeType(), $this->supportedAttributes);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
     }
 
     /**
