+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/Filter/GroupsFilter.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter;
 
-use Doctrine\ODM\MongoDB\Query\Builder as QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\Operators;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldFilterInterface;
 
@@ -13,16 +12,10 @@ use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldFilterInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GroupsFilter implements FieldFilterInterface
+class GroupsFilter extends AbstractFilter implements FieldFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedFields;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the filter
@@ -41,39 +34,9 @@ class GroupsFilter implements FieldFilterInterface
     /**
      * {@inheritdoc}
      */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!($queryBuilder instanceof QueryBuilder)) {
-            throw new \InvalidArgumentException(
-                'Query builder should be an instance of Doctrine\ODM\MongoDB\Query\Builder'
-            );
-        }
-
-        $this->qb = $queryBuilder;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function supportsField($field)
     {
         return in_array($field, $this->supportedFields);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
     }
 
     /**
