+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/Filter/PriceFilter.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter;
 
-use Doctrine\ODM\MongoDB\Query\Builder as QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\ProductQueryUtility;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\Operators;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\AttributeFilterInterface;
@@ -15,16 +14,10 @@ use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class PriceFilter implements AttributeFilterInterface
+class PriceFilter extends AbstractFilter implements AttributeFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedAttributes;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the filter
@@ -43,39 +36,9 @@ class PriceFilter implements AttributeFilterInterface
     /**
      * {@inheritdoc}
      */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!($queryBuilder instanceof QueryBuilder)) {
-            throw new \InvalidArgumentException(
-                'Query builder should be an instance of Doctrine\ODM\MongoDB\Query\Builder'
-            );
-        }
-
-        $this->qb = $queryBuilder;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function supportsAttribute(AttributeInterface $attribute)
     {
         return in_array($attribute->getAttributeType(), $this->supportedAttributes);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
     }
 
     /**
