+++ src/Pim/Bundle/CatalogBundle/Doctrine/MongoDBODM/Filter/StringFilter.php
@@ -2,11 +2,9 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\Filter;
 
-use Doctrine\ODM\MongoDB\Query\Builder as QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\MongoDBODM\ProductQueryUtility;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\Operators;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\AttributeFilterInterface;
-use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldFilterInterface;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 
 /**
@@ -16,57 +14,23 @@ use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class StringFilter implements AttributeFilterInterface, FieldFilterInterface
+class StringFilter extends AbstractFilter implements AttributeFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedAttributes;
-
-    /** @var array */
-    protected $supportedFields;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the filter
      *
      * @param array $supportedAttributes
-     * @param array $supportedFields
      * @param array $supportedOperators
      */
     public function __construct(
         array $supportedAttributes = [],
-        array $supportedFields = [],
         array $supportedOperators = []
     ) {
         $this->supportedAttributes = $supportedAttributes;
-        $this->supportedFields     = $supportedFields;
         $this->supportedOperators  = $supportedOperators;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!($queryBuilder instanceof QueryBuilder)) {
-            throw new \InvalidArgumentException(
-                'Query builder should be an instance of Doctrine\ODM\MongoDB\Query\Builder'
-            );
-        }
-
-        $this->qb = $queryBuilder;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsField($field)
-    {
-        return in_array($field, $this->supportedFields);
     }
 
     /**
@@ -75,22 +39,6 @@ class StringFilter implements AttributeFilterInterface, FieldFilterInterface
     public function supportsAttribute(AttributeInterface $attribute)
     {
         return in_array($attribute->getAttributeType(), $this->supportedAttributes);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
     }
 
     /**
