+++ src/Pim/Bundle/CatalogBundle/Resources/config/query_builders.yml
@@ -135,7 +135,6 @@ services:
         class: %pim_catalog.doctrine.query.filter.string.class%
         arguments:
             - ['pim_catalog_text', 'pim_catalog_textarea']
-            - []
             - ['STARTS WITH', 'ENDS WITH', 'CONTAINS', 'DOES NOT CONTAIN', '=', 'EMPTY']
         tags:
             - { name: 'pim_catalog.doctrine.query.filter', priority: 30 }
@@ -144,7 +143,6 @@ services:
         class: %pim_catalog.doctrine.query.filter.string.class%
         arguments:
             - ['pim_catalog_identifier']
-            - []
             - ['STARTS WITH', 'ENDS WITH', 'CONTAINS', 'DOES NOT CONTAIN', '=', 'IN']
         tags:
             - { name: 'pim_catalog.doctrine.query.filter', priority: 30 }
